Added clauto-web submodule and incorporated web app into release flow

+++ devscripts/deplocal.sh
@@ -7,9 +7,18 @@
 
 # Ensure clautod is installed
 if [ "$(dpkg-query -W --showformat='${Status}\n' clautod | grep 'install ok installed')" == "" ] ; then
-    echo "[deplocal] clautod not installed"
+    echo "[clautod deplocal] clautod not installed"
     exit 1
 fi
+
+# Build clauto-web
+cd clautod/clauto-web
+vue-cli-service build
+if [ $? -ne 0 ] ; then
+    echo "[clautod deplocal] Web build failed. Not deploying."
+    exit 1
+fi
+cd ../..
 
 # Bring down the clautod instance
 sudo systemctl stop clautod
@@ -24,13 +33,13 @@ sudo cp -r clautod/* /usr/share/clauto/clautod
 
 # There may have been config changes, so enact config migration
 if ! sudo /usr/share/clauto/clauto-common/sh/cfgmig.sh clautod ; then
-    echo "[deplocal] Config migration failure detected. Exiting without restarting clautod"
+    echo "[clautod deplocal] Config migration failure detected. Exiting without restarting clautod"
     exit 1
 fi
 
 # There may have been database changes, so enact database migration
 if ! sudo /usr/share/clauto/clautod/dbmig/dbmig.sh ; then
-    echo "[deplocal] DB migration failure detected. Exiting without restarting clautod"
+    echo "[clautod deplocal] DB migration failure detected. Exiting without restarting clautod"
     exit 1
 fi
